+++ cvat-ui/src/utils/event-recorder.ts
